OBJKT Width display on High-density pixel screens
Requesting a PR to allow a unified display of objects across various screen resolutions; while keeping as much information as possible visible underneath the object preview.
Using the a 'vw' derivative for the width, creates a similar layout // look-n-feel, if viewed on either conventional pixel density displays or retina. Currently the object preview appears rather small on retina displays, and it mostly affecting video and static image artworks which are horizontal and specially interactive objects which require the user to navigate/use the artwork. Thank you.