Change price of swapped edition
This patch allow a collector to change the price of swapped editions of an objkt.

Currently, to change a price, you have to unswap, wait for tx to complete, then swap at new price.

This patch do the same thing but in a batch that cancel current swap and do a new swap with new price provided by collector through UI.